refactor accumulator into separate module
